Remove I prefix in interfaces
