--- src/controllers/BaseController.ts
@@ -1,3 +1,0 @@
-export abstract class BaseController {
-    constructor() {/** */}
-}
