+++ src/controllers/HealthController.ts
@@ -1,15 +1,13 @@
 import express from 'express';
 
 import { CatchUnexpected } from '../lib/Decorators';
-import { BaseController } from './BaseController';
 import { HealthService } from '../services/HealthService';
 import { Logger } from '../lib/Logger';
 
 @CatchUnexpected(500)
-export class HealthController extends BaseController {
+export class HealthController {
 
     constructor(private healthService: HealthService, private logger: Logger) {
-        super();
         this.logger.debug('initialized');
     }
