--- src/data/entities/User.ts
@@ -1,8 +1,0 @@
-export interface User {
-    id: number;
-    name: string;
-    surname: string;
-    email: string;
-
-    password: string;
-}
