--- src/services/BaseService.ts
@@ -1,3 +1,0 @@
-export abstract class BaseService {
-    constructor() {/** */}
-}
